Messages & validation done!

+++ database/migrations/2023_03_02_184325_create_trucks_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('maker', 255);
             $table->string('plate', 20);
-            $table->integer('make_year');
+            $table->unsignedSmallInteger('make_year');
             $table->text('mechanic_notices');
             $table->unsignedBigInteger('mechanic_id');
             $table->foreign('mechanic_id')->references('id')->on('mechanics');
